+++ resources/js/signrequestapp/components/modals/Ending.vue
@@ -9,6 +9,11 @@
       <p class="text-white">
         {{'endModal.body' | trans}}
       </p>
+      <b-btn variant="success" block @click="handleCopyClipboard">
+        <b v-html="buttonCopyText">
+
+        </b>
+      </b-btn>
       <b-btn variant="warning" block @click="handleOkEndModal">
         <b>
           {{'endModal.button' | trans}}
@@ -64,6 +69,7 @@
     },
     data: function() {
       return {
+        buttonCopyText: 'Click to Copy',
         review: {
 
           rating: null,
@@ -124,6 +130,15 @@
         this.errors.rating.state = null
         this.errors.rating.data = ''
       },
+      handleCopyClipboard() {
+        let host = window.location.host
+        let pathname = '/sign-request/sign-file/'
+        this.copyTextToClipboard(host + pathname + this.$store.state.signerToken)
+        this.buttonCopyText = '<i class="fa fa-check"></i> Copied!'
+        setTimeout(() => {
+          this.buttonCopyText = "Click to Copy"
+        }, 500);
+      },
       sendReview() {
         let locale = this.$lang.getLocale()
 
@@ -156,6 +171,41 @@
             }
 
           })
+      },
+      fallbackCopyTextToClipboard(text) {
+        var textArea = document.createElement("textarea");
+        textArea.value = text;
+
+        // Avoid scrolling to bottom
+        textArea.style.top = "0";
+        textArea.style.left = "0";
+        textArea.style.position = "fixed";
+
+        document.body.appendChild(textArea);
+        textArea.focus();
+        textArea.select();
+
+        try {
+          var successful = document.execCommand('copy');
+          var msg = successful ? 'successful' : 'unsuccessful';
+          console.log('Fallback: Copying text command was ' + msg);
+        } catch (err) {
+          console.error('Fallback: Oops, unable to copy', err);
+        }
+
+        document.body.removeChild(textArea);
+      },
+
+      copyTextToClipboard(text) {
+        if (!navigator.clipboard) {
+          fallbackCopyTextToClipboard(text);
+          return;
+        }
+        navigator.clipboard.writeText(text).then(function() {
+          console.log('Async: Copying to clipboard was successful!');
+        }, function(err) {
+          console.error('Async: Could not copy text: ', err);
+        });
       }
 
     },