+++ resources/js/signrequestapp/i18n/en/endModal.js
@@ -1,7 +1,7 @@
 export default {
   'en.endModal': {
     'title': 'Thank You For Using This Apps!',
-    'body': 'Your file is successfully stored and secured with encryption AES-256, soon your signer will received your sign request by email. Please check your email.',
+    'body': 'Your file is successfully stored and secured with encryption AES-256, soon your signer will received your sign request by email, so do you. Or, you can share your sign request link by click the button below ',
     'button': 'Sign request another file',
     'share': {
       'body': 'If you feel this apps is helpful, please share by click one of these buttons bellow.'